refactor: moved toolbar preview code to utils
